Provide a tutorial
Resolves https://github.com/NickMcConnell/NarSil/issues/1 .
The contents of the tutorial tries to remain faithful to the one in Sil 1.3.  The layout of the datafile defining the tutorial and the source code supporting that derive from a draft tutorial for Angband 4.2.
The tutorial is accessible via the menus for the Windows and Mac front ends.  For front ends using main.c, an option to load the tutorial is included in the savefile selection list generated when the "-c" option is used.
Expands the attributes stored in textblocks from uint8_t to int so that is possible to display solid or hybrid walls within a textblock as they would appear in game.
Extends expressions so the expression's result can be provided by from an integer set at run time in lieu of calling a function.  That's to allow the pvals for staves placed in the tutorial to depend on the number of staves in the stack.

+++ src/mon-list.c
@@ -302,7 +302,7 @@ void monster_list_sort(monster_list_t *list,
  * \param entry is the monster list entry to display.
  * \return a color for the monster entry.
  */
-uint8_t monster_list_entry_line_color(const monster_list_entry_t *entry)
+int monster_list_entry_line_color(const monster_list_entry_t *entry)
 {
 	/* Display uniques in a special colour */
 	if (rf_has(entry->race->flags, RF_UNIQUE))
