+++ src/mon-list.h
@@ -57,6 +57,6 @@ int monster_list_standard_compare(const void *a, const void *b);
 int monster_list_compare_exp(const void *a, const void *b);
 void monster_list_sort(monster_list_t *list,
 					   int (*compare)(const void *, const void *));
-uint8_t monster_list_entry_line_color(const monster_list_entry_t *entry);
+int monster_list_entry_line_color(const monster_list_entry_t *entry);
 
 #endif /* MONSTER_LIST_H */
