+++ src/ui-mon-list.c
@@ -98,7 +98,7 @@ static void monster_list_format_section(const monster_list_t *list, textblock *t
 	for (index = 0; index < total && line_count < lines_to_display; index++) {
 		char alert[20] = { '\0' };
 		char location[20] = { '\0' };
-		uint8_t line_attr;
+		int line_attr;
 		size_t full_width;
 		size_t name_width;
 		uint16_t count_in_section = 0;
