+++ src/z-textblock.h
@@ -30,12 +30,12 @@ void textblock_free(textblock *tb);
 
 
 void textblock_append(textblock *tb, const char *fmt, ...);
-void textblock_append_c(textblock *tb, uint8_t attr, const char *fmt, ...);
-void textblock_append_pict(textblock *tb, uint8_t attr, int c);
+void textblock_append_c(textblock *tb, int attr, const char *fmt, ...);
+void textblock_append_pict(textblock *tb, int attr, int c);
 void textblock_append_textblock(textblock *tb, const textblock *tba);
 
 const wchar_t *textblock_text(textblock *tb);
-const uint8_t *textblock_attrs(textblock *tb);
+const int *textblock_attrs(textblock *tb);
 
 size_t textblock_calculate_lines(textblock *tb, size_t **line_starts,
 								 size_t **line_lengths, size_t width);
@@ -43,14 +43,14 @@ size_t textblock_calculate_lines(textblock *tb, size_t **line_starts,
 void textblock_to_file(textblock *tb, ang_file *f, int indent, int wrap_at);
 
 extern ang_file *text_out_file;
-extern void (*text_out_hook)(uint8_t a, const char *str);
+extern void (*text_out_hook)(int a, const char *str);
 extern int text_out_wrap;
 extern int text_out_indent;
 extern int text_out_pad;
 
-extern void text_out_to_file(uint8_t attr, const char *str);
+extern void text_out_to_file(int attr, const char *str);
 extern void text_out(const char *fmt, ...);
-extern void text_out_c(uint8_t a, const char *fmt, ...);
+extern void text_out_c(int a, const char *fmt, ...);
 extern void text_out_e(const char *fmt, ...);
 
 typedef void (*text_writer)(ang_file *f);
